Moved support classes into own private header
